% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_Cpk.R
\name{chart_Cpk}
\alias{chart_Cpk}
\title{Create a capability chart for statistical process control}
\usage{
chart_Cpk(data)
}
\arguments{
\item{data}{}

\item{factory}{}
}
\value{

}
\description{
Generate a histogram type chart from a set of consecutive measurements.
}
\details{
This type of chart is typically applied in product manufacturing to monitor
deviations from the target value over time. It is usually accompanied by 
the statistical process control time series \code{\link{chart_I}} and 
\code{\link{chart_IMR}}
}
\examples{


}
\references{
For a complete case study application refer to \url{https://j-ramalho.github.io/industRial/}
}
