% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_I.R
\name{chart_I}
\alias{chart_I}
\title{Create IMR chart for statistical process control}
\usage{
chart_I(data)
}
\arguments{
\item{data}{A dataset generated by the function \code{\link{process_stats}}}
}
\value{
This function returns an object of class ggplot
}
\description{
Generate a single point time series chart from a set of consecutive measurements.
}
\details{
This type of chart is typically applied in product manufacturing to monitor
deviations from the target value over time. It is usually accompanied by 
the \code{\link{chart_IMR}}
}
\references{
For a complete case study application refer to \url{https://j-ramalho.github.io/industRial/}
}
