% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_R.R
\name{chart_R}
\alias{chart_R}
\title{Create R chart for statistical process control}
\usage{
chart_R(data, factory)
}
\arguments{
\item{data}{}

\item{factory}{}
}
\value{

}
\description{
Generate a ranges chart from a set of consecutive measurements by calculating
their mean on each time point.
}
\details{
This type of chart is typically applied in product manufacturing to monitor
deviations from the target value over time. It is usually accompanied by 
the \code{\link{chart_xbar}}
}
\references{
For a complete case study application refer to \url{https://j-ramalho.github.io/industRial/}
}
