% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_xbar.R
\name{chart_xbar}
\alias{chart_xbar}
\title{Create xbar chart for statistical process control}
\usage{
chart_xbar(data, factory)
}
\arguments{
\item{data}{}

\item{factory}{}
}
\value{

}
\description{
Generate an means chart from a set of consecutive measurements by calculating
their mean on each time point.
}
\details{
This type of chart is typically applied in product manufacturing to monitor
deviations from the target value over time. It is usually accompanied by 
the \code{\link{chart_xbarR}}
}
\references{
For a complete case study application refer to \url{https://j-ramalho.github.io/industRial/}
}
