% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_SPCdata.R
\name{clean_SPCdata}
\alias{clean_SPCdata}
\title{Clean control data for Statistical Process Control Charts (SPC)}
\usage{
clean_SPCdata(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
This function takes a dataset consisting of consecutive measurements and
cleans it so that it can be used in SPC charts.
}
\examples{

}
\references{
For a complete case study application refer to \url{https://j-ramalho.github.io/industRial/}
}
