% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_dial_control.R
\docType{data}
\name{dial_control}
\alias{dial_control}
\title{Collection of visual defects on watch dial production.}
\format{
An object of class tibble with 58 observations on 4 variables.
}
\source{
Original data set.
}
\usage{
dial_control
}
\description{
This data set contains observations of visual defects present
in watch dials such as indentations and scratches taken during production.
It provides a practical case to establish pareto charts typically with a 
function like \code{\link[qicharts2]{paretochart}}.

\describe{
  \item{Operator}{The shop floor operator collecting the data}
  \item{Date}{Data collection date}
  \item{Defect}{Defect type ("Indent", "Scratch")}
  \item{Location}{Position on the watch dial refered to as the hour (1h, 2h)}
  \item{id}{Part unique id number}
}
}
\examples{
head(dial_control)
}
\references{
For a complete case study application refer to \url{https://j-ramalho.github.io/industRial/}.
}
\keyword{datasets}
