% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ebike_hardening.R
\docType{data}
\name{ebike_hardening}
\alias{ebike_hardening}
\alias{ebike_hardening2}
\title{Cycles to failure of ebikes frames after temperature treatment.}
\format{
A tibble with 4 observations on 6 variables.
}
\source{
Original data set.
}
\usage{
ebike_hardening
}
\description{
A data set with the results of aging tests on several groups
of ebikes frames (g1, g2, ...). Each entry corresponds to 
the number of cycles to failure for each level of treatment temperature-

\describe{
  \item{temperature}{Position of the part on the device}
  \item{g1}{group 1, remaining groups have names g2 to g5}
}
}
\details{
The ebike_hardening2 dataset contains alternative data that gives
non significant results in the analysis of variance study.
}
\examples{
data(ebike_hardening)
}
\references{
For a complete case study application refer to \url{https://j-ramalho.github.io/industRial/}.
}
\keyword{datasets}
