% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_formula.R
\name{expand_formula}
\alias{expand_formula}
\title{Formula expansion}
\usage{
expand_formula(formulae)
}
\arguments{
\item{formulae}{Takes as input object of class formula, e.g.: Y ~ A * B, see ?formula for 
syntax details}
}
\value{
Returns a character vector such as A + B + A:B
}
\description{
Takes a linear model formula and returns it expanded version.
}
\details{
Supports verification and understanding of the creation of linear models 
syntax such as *,+ and other conventions.
}
\references{
For an example application refer to \url{https://j-ramalho.github.io/industRial/}
}
