% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_expansion.R
\name{formula_expansion}
\alias{formula_expansion}
\title{Formula expansion}
\usage{
formula_expansion(formula)
}
\arguments{
\item{formula}{}
}
\value{
Returns a character vector such as A + B + A:B
}
\description{
Takes a linear model formula and returns it expanded version.
}
\details{
This function supports verification and understanding of the creation of 
linear models syntax such as *,+ and other conventions.
}
\references{
For an example application refer to \url{https://j-ramalho.github.io/industRial/}
}
