% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_perfume_experiment.R
\docType{data}
\name{perfume_experiment}
\alias{perfume_experiment}
\title{Correlation matrix of the input variables of an experiment design in 
perfume formulation.}
\format{
A tibble with 22 observations on 23 variables.
}
\source{
Original data set.
}
\usage{
perfume_experiment
}
\description{
The data set contains the expected correlation 
(expressed in 1 to 10) of an experiment anonymized input variables. The 
dataset consists in a double entry table with the same variables in row
and column. It is coded as a tibble but subsequent utilization in network
plots requires it to be converted to a matrix format.
}
\examples{
data(perfume_experiment)

}
\references{
For a complete case study application refer to \url{https://j-ramalho.github.io/industRial/}.
}
\keyword{datasets}
