% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_pet_doe.R
\docType{data}
\name{pet_doe}
\alias{pet_doe}
\title{A factorial design for the improvement of PET film tensile strength.}
\format{
An object of classes design and data.frame with 12 observations of 
4 variables.
}
\source{
Original data set generated with the function 
\code{\link[DoE.base]{fac.design}} form the package DoE.base.
}
\usage{
pet_doe
}
\description{
The data corresponds to full factorial design with two factors
coded as +/- and 3 replicates for each combination.

\describe{
  \item{A}{PET formulation A (factor)}
  \item{B}{PET formulation B (factor)}
  \item{replicate}{the measurement replicate I to III (factor)}
  \item{yield}{the output variable measured on the PET, (numerical)}
}
}
\examples{
data(pet_doe)
contrasts(pet_doe$A)

}
\references{
For a complete case study application refer to \url{https://j-ramalho.github.io/industRial/}
}
\keyword{datasets}
