% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_Cpk.R
\name{process_Cpk}
\alias{process_Cpk}
\title{Calculate process capability index for Statistical Process Control}
\usage{
process_Cpk(UCL, LCL, mean, sd)
}
\arguments{
\item{UCL}{the process upper control limit}

\item{LCL}{the process lower control limit}

\item{mean}{the process mean}

\item{sd}{the process standard deviation}
}
\value{
This function returns an object of class numeric
}
\description{
This function takes process variables and calculates the Cpk index which 
is a measure of the process centering and variability against specification.
}
\examples{
process_Cpk(100, 0, 10, 3)
}
\references{
For a complete case study application refer to \url{https://j-ramalho.github.io/industRial/}
}
