% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_stats.R
\name{process_stats}
\alias{process_stats}
\title{Calculate summary statistics for Statistical Process Control}
\usage{
process_stats(data, part_spec_percent)
}
\arguments{
\item{data}{This function takes the dataset tablet_thickness cleaned with the clean_names function from 
the {janitor} package}

\item{part_spec_percent}{the process tolerance in percentage.}
}
\value{
This function returns an object with class tibble (tbl_df)
}
\description{
This function takes process variables and calculates summary statistics and 
presents them in a easy readable table format.
}
\references{
For a complete case study application refer to \url{https://j-ramalho.github.io/industRial/}
}
