% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_solarcell_fill.R
\docType{data}
\name{solarcell_fill}
\alias{solarcell_fill}
\title{Yearly outputs and fills factor of solarcells of different types.}
\format{
A tibble with 15 observations of 3 variables.
}
\source{
Original data set.
}
\usage{
solarcell_fill
}
\description{
A dataset with the energy output resulting from tests on
solarcells made of three different configurations. The fill factor 
provides an indication of the cell quality and is a non controlled variable
that can be taken into consideration in an analysis of covariance to better
assess the output variation from material to material.

\describe{
  \item{material}{The solar cell material (character)}
  \item{output}{he yearly energy output (numberic)}
  \item{fillfactor}{The fill factor measured for each cell (numberic)}
}
}
\examples{
hist(solarcell_fill$output)
}
\references{
For a complete case study application refer to \url{https://j-ramalho.github.io/industRial/}.
}
\keyword{datasets}
