% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_industRial.R
\name{theme_industRial}
\alias{theme_industRial}
\title{Custom theme "industRial" for the book industRial Data Science plots}
\usage{
theme_industRial(
  font_size = 14,
  font_family = "",
  line_size = 0.5,
  rel_small = 12/14,
  rel_tiny = 11/14,
  rel_large = 16/14
)
}
\arguments{
\item{base_size}{}

\item{base_family}{}
}
\value{

}
\description{
This theme aims at optimal balance between readability and precision. It has 
adapted from the package {cowplot} by Claus O.Wilke and reflects the principles
of his book \href{https://clauswilke.com/dataviz/}{Fundamentals of Data Visualization }
}
\details{
Apply this theme by adding it at the end of the code of any \code{\link[ggplot2]{ggplot}} chart.
It basically combines the half open theme with a grid background from {cowplot}
}
\references{
For a complete case study application refer to \url{https://j-ramalho.github.io/industRial/}
}
