% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbar_plot.R
\name{xbar_plot}
\alias{xbar_plot}
\title{Create means plot for statistical process control}
\usage{
xbar_plot(data, factory)
}
\arguments{
\item{data}{}

\item{factory}{}
}
\value{

}
\description{
Generate an xbar plot from a set of consecutive measurements by calculating
their mean on each timepoint.
}
\details{
This type of plot is typically applied in product manufacturing to monitor
deviations from the target value over time.
}
