% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_IMR.R
\name{chart_IMR}
\alias{chart_IMR}
\title{Create R MR chart for statistical process control}
\usage{
chart_IMR(data, factory)
}
\arguments{
\item{data}{}

\item{factory}{}
}
\value{

}
\description{
Generate a moving range chart chart from a set of consecutive measurements.
}
\details{
This type of chart is typically applied in product manufacturing to monitor
deviations from the target value over time. It is usually accompanied by 
the \code{\link{chart_IMR}}
}
\references{
For a complete case study application refer to \url{https://j-ramalho.github.io/industRial/}
}
