% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/off_spec.R
\name{off_spec}
\alias{off_spec}
\title{Calculate percentage of out of specification for Statistical Process Control}
\usage{
off_spec(UCL, LCL, mean, sd)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
This function takes process variables and calculates the probability that
parts are produced out of specification on the long run.
}
\examples{

}
\references{
For a complete case study application refer to \url{https://j-ramalho.github.io/industRial/}
}
