% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_petdelivery.R
\docType{data}
\name{pet_delivery}
\alias{pet_delivery}
\title{Tensile strength values on PET raw material for the clothing industry}
\format{
A tibble with 28 observations on 2 variables
}
\source{
Original data set
}
\usage{
pet_delivery
}
\description{
\describe{
  \item{A}{Tensile strenght measurements for product A [Mpa] (numeric)}
  \item{B}{Tensile strenght measurements for product B [Mpa] (numeric)}
}
}
\references{
For a complete case study application refer to \url{https://j-ramalho.github.io/industRial/}
}
\keyword{datasets}
