% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_Cpk.R
\name{process_Cpk}
\alias{process_Cpk}
\title{Calculate process capability index for Statistical Process Control}
\usage{
process_Cpk(UCL, LCL, mean, sd)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
This function takes process variables and calculates the Cpk index which 
is a measure of the process centering and variability against specification.
}
\examples{

}
\references{
For a complete case study application refer to \url{https://j-ramalho.github.io/industRial/}
}
