% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_stats.R
\name{process_stats}
\alias{process_stats}
\title{Calculate summary statistics for Statistical Process Control}
\usage{
process_stats(data, pod_spec_percent)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
This function takes process variables and calculates summary statistics and 
presents them in a easy readable table format.
}
\examples{

}
\references{
For a complete case study application refer to \url{https://j-ramalho.github.io/industRial/}
}
