% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_stats_table.R
\name{process_stats_table}
\alias{process_stats_table}
\title{Summary statistics table outputs for Statistical Process Control}
\usage{
process_stats_table(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
This function takes summary statistics and presents them in a easy readable 
table format.
}
\examples{

}
\references{
For a complete case study application refer to \url{https://j-ramalho.github.io/industRial/}
}
